packages/cli-py/pyproject.toml
[project]
name = "agent-task"
version = "0.1.0"
description = "A portable task contract format"
authors = [{name = "Shail Shah"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "jsonschema>=4.20.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0"
]

[project.scripts]
agent-task = "agent_task.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88

[tool.mypy]
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
